Redirect invocations on Piple to funcs in binding: fixes #8
@a0x77n this is the hack we discuss, which fixes #8
beautifully. Essentially, all calls made on the GremlinPipeline are
redirected to functions in the binding with the specified name. Also,
if no such function exists, the handler is not called again, as we're
not trying to invoke the method on a GremlinPipeline object.

I've also added an example step and tested that it can be overwritten
at runtime.