Use noTX in instructionlinker.
Apparently, once we disable the WriteAheadLog (WAL), we cannot reenable it without restarting the server, and that means, once we've used the importer plugin, we can no longer use tx-connections to the database.

For now, I have changed the instructionlinker code to use noTX, which means we should consider the database locked while the instructionlinker is running. On the positive side, this makes edge creation a lot faster.

On the long run, we should find out whether there is a way to reenable WAL. My guess is that this is an orientdb bug. It's more important right now though that the server never has to be restarted.